First 3.12 build
First working build, and a lot has changed.

Converted each of the pages folders into Packages so that they are a easier to work with.
Each package has a __init__ that sets up the Blueprint, and  a routes.py that gives the page its functionality.

updated all the requirements.text to the latest packages that are released.